refactor: rename config module to settings, update imports across the codebase
